+++ Clicked/Core/Utils.lua
@@ -1,5 +1,14 @@
 local L = LibStub("AceLocale-3.0"):GetLocale("Clicked")
 
+local KEYBIND_ORDER_LIST = {
+	"BUTTON1", "BUTTON2", "BUTTON3", "BUTTON4", "BUTTON5", "MOUSEWHEELUP", "MOUSEWHEELDOWN",
+	"`", "0", "1", "2", "3", "4", "5", "6", "7", "8", "9", "-", "=",
+	"NUMPAD0", "NUMPAD1", "NUMPAD2", "NUMPAD3", "NUMPAD4", "NUMPAD5", "NUMPAD6", "NUMPAD7", "NUMPAD8", "NUMPAD9", "NUMPADDIVIDE", "NUMPADMULTIPLY", "NUMPADMINUS", "NUMPADPLUS", "NUMPADDECIMAL",
+	"F1", "F2", "F3", "F4", "F5", "F6", "F7", "F8", "F9", "F10", "F11", "F12", "F13",
+	"A", "B", "C", "D", "E", "F", "G", "H", "I", "J", "K", "L", "M", "N", "O", "P", "Q", "R", "S", "T", "U", "V", "W", "X", "Y", "Z",
+	"TAB", "CAPSLOCK", "INSERT", "DELETE", "HOME", "END", "PAGEUP", "PAGEDOWN", "[", "]", "\\", ";", "'", ",", ".", "/"
+}
+
 -- /run local a,b,c=table.concat,{},{};for d=1,GetNumShapeshiftForms() do local _,_,_,f=GetShapeshiftFormInfo(d);local e=GetSpellInfo(f);b[#b+1]=e;c[#c+1]=f;end print("{ "..a(c, ",").." }, --"..a(b,", "))
 local shapeshiftForms = {
 	-- Arms Warrior
@@ -121,6 +130,54 @@ local function safecall(func, ...)
 	if func then
 		return xpcall(func, errorhandler, ...)
 	end
+end
+
+local function GetKeybindIndex(keybind)
+	local mods = {}
+	local result = ""
+
+	for match in string.gmatch(keybind, "[^-]+") do
+		table.insert(mods, match)
+		result = match
+	end
+
+	table.remove(mods, #mods)
+
+	local index = #KEYBIND_ORDER_LIST + 1
+	local found = false
+
+	for i = 1, #KEYBIND_ORDER_LIST do
+		if KEYBIND_ORDER_LIST[i] == result then
+			index = i
+			found = true
+			break
+		end
+	end
+
+	-- register this unknown keybind for this session
+	if not found then
+		table.insert(KEYBIND_ORDER_LIST, result)
+	end
+
+	for i = 1, #mods do
+		if mods[i] == "CTRL" then
+			index = index + 1000
+		end
+
+		if mods[i] == "ALT" then
+			index = index + 10000
+		end
+
+		if mods[i] == "SHIFT" then
+			index = index + 100000
+		end
+
+		if mods[i] == "META" then
+			index = index + 1000000
+		end
+	end
+
+	return index
 end
 
 StaticPopupDialogs["CLICKED_INCOMPATIBLE_ADDON"] = {
@@ -339,6 +396,59 @@ function Clicked:IsStringNilOrEmpty(string)
 
 	string = tostring(string)
 	return #string == 0
+end
+
+--- Compare two bindings, for use in a comparison function such as `table.sort`
+--- This function is stable and will return the opposite result if called with inverted parameters
+---
+--- @param left table
+--- @param right table
+--- @return boolean
+function Clicked:CompareBindings(left, right)
+	assert(type(left) == "table", "bad argument #1, expected table but got " .. type(left))
+	assert(type(right) == "table", "bad argument #2, expected table but got " .. type(right))
+
+	do
+		local leftLoad = Clicked:CanBindingLoad(left)
+		local rightLoad = Clicked:CanBindingLoad(right)
+
+		if leftLoad and not rightLoad then
+			return true
+		end
+
+		if not leftLoad and rightLoad then
+			return false
+		end
+	end
+
+	if left.keybind == "" and right.keybind ~= "" then
+		return false
+	end
+
+	if left.keybind ~= "" and right.keybind == "" then
+		return true
+	end
+
+	if left.keybind == right.keybind then
+		local leftValue = self:GetActiveBindingValue(left)
+		local rightValue = self:GetActiveBindingValue(right)
+
+		if leftValue ~= nil and rightValue == nil then
+			return true
+		end
+
+		if leftValue == nil and rightValue ~= nil then
+			return false
+		end
+
+		if leftValue == nil and rightValue == nil then
+			return left.identifier < right.identifier
+		end
+
+		return leftValue < rightValue
+	end
+
+	return GetKeybindIndex(left.keybind) < GetKeybindIndex(right.keybind)
 end
 
 --- Get all available shapeshift forms the the specified spec ID.