+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
+## [Unreleased]
+
+## Added
+
+- Add an option to show unit-frame bindings on tooltips ([h2oboi89](https://github.com/h2oboi89))
+
 ## [0.15.5] - 2021-02-15
 
 ### Fixed